+++ src/p_enemy.c
@@ -1899,9 +1899,9 @@ void A_Explode(mobj_t *actor, player_t *player, pspdef_t *psp)
         {
             barrelms = I_GetTimeMS() + BARRELMS;
 
-            if (gp_vibrate_barrels)
+            if (gp_rumble_barrels)
             {
-                I_GamepadVibration(20000 * gp_vibrate_barrels / 100);
+                I_GamepadVibration(20000 * gp_rumble_barrels / 100);
                 barrelvibrationtics = TICRATE;
             }
         }