+++ src/p_inter.c
@@ -2322,9 +2322,9 @@ void P_DamageMobj(mobj_t *target, mobj_t *inflicter, mobj_t *source, int damage,
         if (r_shake_damage)
             I_UpdateBlitFunc(tplayer->damagecount);
 
-        if (gp_vibrate_damage)
+        if (gp_rumble_damage)
         {
-            I_GamepadVibration((30000 + (100 - MIN(tplayer->health, 100)) / 100 * 30000) * gp_vibrate_damage / 100);
+            I_GamepadVibration((30000 + (100 - MIN(tplayer->health, 100)) / 100 * 30000) * gp_rumble_damage / 100);
             damagevibrationtics += BETWEEN(12, damage, 100);
         }